Bug fixing
git-svn-id: https://svn.kuali.org/repos/student/contrib/ksa/trunk@30489 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ksa-core/src/main/java/com/sigmasys/kuali/ksa/gwt/client/model/AccountModel.java
@@ -99,11 +99,11 @@ public class AccountModel extends AbstractModel {
         set(COLUMN_IS_KIM_ACCOUNT, kimAccount);
     }
 
-    public BigDecimal getCreditLimit() {
+    public Double getCreditLimit() {
         return get(COLUMN_CREDIT_LIMIT);
     }
 
-    public void setCreditLimit(BigDecimal creditLimit) {
+    public void setCreditLimit(Double creditLimit) {
         set(COLUMN_CREDIT_LIMIT, creditLimit);
     }
